[PC-19443] Deploy GCE from Airflow

+++ algo_training/deploy_model.py
@@ -233,7 +233,7 @@ def main(
         min_nodes=int(min_nodes),
         max_nodes=int(max_nodes),
         machine_type=machine_type,
-        traffic_percentage=float(traffic_percentage),
+        traffic_percentage=int(traffic_percentage),
     )
     handler = ModelHandler(region, GCP_PROJECT_ID, model_params, endpoint_params)
     # Upload new model to registery
